hf_rag_project/requirements.txt
datasets
transformers
torch
sentence-transformers
faiss-cpu
trulens-eval
openai
python-dotenv
